% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strange_attractor.R
\name{attractor_generic}
\alias{attractor_generic}
\title{Generate strange attractor.}
\usage{
attractor_generic(a, n, x0 = 1, y0 = 1, dims = c(600, 600),
  progress = (n >= 1e+05), n_discretize = 1e+06, qs = qs,
  discretize = TRUE)
}
\arguments{
\item{a}{Numeric vector of length 14}

\item{n}{Number of points to generate}

\item{x0}{Initial value for x}

\item{y0}{Initial value for y}

\item{dims}{Dimensions of resulting matrix}

\item{progress}{If TRUE, displays a progress bar}

\item{n_discretize}{Number of values to generate before computing the range}

\item{qs}{quantile cutoff}

\item{discretize}{If \code{TRUE} discretizes to a matrix with dimensions \code{dims} else a data frame}
}
\value{
if \code{discretize == TRUE} a matrix with dimensions \code{dims}, else a data frame with columns \code{x} and \code{y}, and \code{n} rows.
}
\description{
Generate strange attractor.
}
\keyword{Internal}
