% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strange_attractor.R
\name{seed_sprott_7e}
\alias{seed_sprott_7e}
\title{Generate attractor for the Sprott 7e attractor.}
\usage{
seed_sprott_7e(fast_forward, seed = if (!missing(fast_forward)) 1 else
  NA)
}
\arguments{
\item{fast_forward}{"Index value" to retrieve}

\item{seed}{Original seed value, passed to \code{\link[=set.seed]{set.seed()}}}
}
\description{
Generate attractor for the Sprott 7e attractor.
}
