% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strange_attractor.R
\name{seeded_attractor}
\alias{seeded_attractor}
\title{Generate attractor from specific seed value}
\usage{
seeded_attractor(i, n = 1e+05, seed = 1)
}
\arguments{
\item{i}{"Index value" to retrieve}

\item{n}{Number of points to generate}

\item{seed}{Original seed value, passed to \code{\link[=set.seed]{set.seed()}}}
}
\description{
Generate attractor from specific seed value
}
