% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strange_attractor.R
\name{strange_attractor}
\alias{strange_attractor}
\title{Generate strange attractor.}
\usage{
strange_attractor(a, n, x0 = 1, y0 = 1, dims = c(600, 600),
  progress = (n >= 1e+05), n_discretize = 1e+06, qs = 0.05,
  discretize = TRUE)
}
\arguments{
\item{a}{Numeric vector of length 14}

\item{n}{Number of points to generate}

\item{x0}{Initial value for x}

\item{y0}{Initial value for y}

\item{dims}{Dimensions of resulting matrix}

\item{progress}{If TRUE, displays a progress bar}

\item{n_discretize}{Number of values to generate before computing the range}

\item{qs}{quantile cutoff}

\item{discretize}{If \code{TRUE} discretizes to a matrix with dimensions \code{dims} else a data frame}
}
\value{
if \code{discretize == TRUE} a matrix with dimensions \code{dims}, else a data frame with columns \code{x} and \code{y}, and \code{n} rows.
}
\description{
Based on Equation 7E of Sprott.
}
\details{
\deqn{x_{i+1} = a_{1} + a_{2} * x_{i} +  a_{3} * y_{i} +  a_{4} * |x_{i}|^{a5}  +  a6 * |y_{i}|^{a7}}

\deqn{y_{i+1} = a_{8} + a_{9} * x_{i} +  a_{10} * y_{i} +  a_{11} * |x_{i}|^{a12}  +  a13 * |y_{i}|^{a14}}
}
\references{
Julien C. Sprott, "Strange Attractors: Creating Patterns in Chaos", page 418, Equation 7e, http://sprott.physics.wisc.edu/fractals/booktext/sabook.pdf
}
